gộp enum dạng status
