--- Domain/Enum/AccountStatus.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Domain.Enum
-{
-    public enum AccountStatus
-    {
-        DEACTIVATED,
-        ACTIVE,
-        SUSPENDED,
-        BANNED
-        
-    }
-}
