--- Domain/Enum/CropStatus.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Domain.Enum
-{
-    public enum CropStatus
-    {
-        ACTIVE,
-        INACTIVE,
-        IN_STOCK // đang được lên product
-    }
-}
