--- Domain/Enum/DiseaseStatus.cs
@@ -1,24 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Domain.Enum
-{
-    public enum DiseaseStatus
-    {
-        DownyMildew,          // Nấm phấn trắng
-        PowderyMildew,        // Nấm phấn
-        LeafSpot,             // Đốm lá
-        BacterialSoftRot,     // Thối nhũn do vi khuẩn
-        FusariumWilt,         // Héo vàng Fusarium
-        Anthracnose,          // Thán thư
-        DampingOff,           // Chết cây con
-        BlackRot,             // Thối đen
-        MosaicVirus,          // Virus khảm
-        AphidInfestation,     // Rệp
-        ThripsDamage,         // Bọ trĩ
-        WhiteflyInfestation   // Ruồi trắng
-    }
-}
