--- Domain/Enum/ProductStatus.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Domain.Enum
-{
-    public enum ProductStatus
-    {
-        DEACTIVED,
-        ACTIVE,
-        //OUT_OF_STOCK,
-        //DISCONTINUED
-    }
-}
